+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     So as of version 0.8, we now have `head x == msb y`.
     So converting for `Vec`tors of `Bit`s to `BitVector`s is no longer index-preserving.
   * Add [QuickCheck](http://hackage.haskell.org/package/QuickCheck) 'Arbitary' and 'CoArbitary' instances for all data types
+  * Make the (Bit)Vector argument the _last_ argument for the following functions: `slice`, `setSlice`, `replaceBit`, `replace`.
 
 ## 0.7.5 **May 7th 2015**
 * New features: